requirements.txt
fastapi
uvicorn
pandas
numpy
openai
python-dotenv
sentence-transformers
faiss-cpu
matplotlib
